requirements.txt
dj-database-url==0.5.0
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
gunicorn==19.8.1
psycopg2==2.7.4
whitenoise==3.3.1
